raspberry_pi/requirements.txt
websockets>=11.0
loguru>=0.7.0
pyserial>=3.5
joblib>=1.3.0
